TINY-11783: Bump `fast-check` major version and fix API references.

+++ modules/agar/src/main/ts/ephox/agar/arbitrary/WeightedChoice.ts
@@ -37,7 +37,7 @@ const choose = <T extends WeightedItem>(candidates: T[]): WeightedList<T> => {
 };
 
 const gChoose = <T extends WeightedItem>(weighted: WeightedList<T>): fc.Arbitrary<Optional<T & AccWeightItem>> =>
-  fc.float({ min: 0, max: weighted.total }).map((w): Optional<T & AccWeightItem> => {
+  fc.float({ min: 0, max: Math.fround(weighted.total) }).map((w): Optional<T & AccWeightItem> => {
     const raw = Arr.find(weighted.list, (d) =>
       w <= d.accWeight
     );
